// src/lib/data/types.ts
export interface Evento {
  id: number;
  provincia: string;
  canton: string;
  parroquia: string;
  comunidad: string;
  lat: number;
  lng: number;
  tipoEvento: string;
  causa: string;
  categoriaEvento: string;
  fechaEvento: string;
  año: number;
  mes: number;
  dia: number;
  horaEvento: string;
  descripcion: string;
  heridos: number;
  desaparecidos: number;
  familiasAfectadas: number;
  afectadosDirectos: number;
  afectadosIndirectos: number;
  familiasDamnificadas: number;
  personasDamnificadas: number;
  personasEvacuadas: number;
  personasAlbergadas: number;
  personasAcogiFamilias: number;
  personasAcogiOtros: number;
  personasEnEvacuacion: number;
  personasResistenEvacuacion: number;
  viviendasAfectadas: number;
  viviendasDestruidas: number;
  educativosAfectados: number;
  educativosDestruidos: number;
  saludAfectados: number;
  saludDestruidos: number;
  puentesAfectados: number;
  puentesDestruidos: number;
  bienesPublicosAfectados: number;
  bienesPublicosDestruidos: number;
  bienesPrivadosAfectados: number;
  bienesPrivadosDestruidos: number;
  viasAfectadasM: number;
  cultivosAfectadosHa: number;
  cultivosPerdidosHa: number;
  coberturaVegetalHa: number;
  animalesAfectados: number;
  animalesMuertos: number;
  fuente: string;
  acciones: string;
  estado: string;
  totalPersonasAfectadas: number;
  muertosDesaparecidos: number;
  nivelEvento: string;
  totalImpactados: number;
  viasAfectadasKm: number;
}